+++ src/silo/query_engine/filter/operators/is_in_covered_region.h
@@ -24,16 +24,16 @@ class IsInCoveredRegion : public Operator {
    enum class Comparator : uint8_t { COVERED, NOT_COVERED };
 
   private:
-   const std::vector<std::pair<size_t, size_t>>* covered_region_ranges;
-   const std::map<size_t, roaring::Roaring>* covered_region_bitmaps;
+   const std::vector<std::pair<uint32_t, uint32_t>>* covered_region_ranges;
+   const std::map<uint32_t, roaring::Roaring>* covered_region_bitmaps;
    uint32_t row_count;
    Comparator comparator;
    uint32_t value;
 
   public:
    explicit IsInCoveredRegion(
-      const std::vector<std::pair<size_t, size_t>>* covered_region_ranges,
-      const std::map<size_t, roaring::Roaring>* covered_region_bitmaps,
+      const std::vector<std::pair<uint32_t, uint32_t>>* covered_region_ranges,
+      const std::map<uint32_t, roaring::Roaring>* covered_region_bitmaps,
       uint32_t row_count,
       Comparator comparator,
       uint32_t value
